PM / devfreq: bw_hwmon: Boost when the screen is turned on
A low bandwidth vote when the screen is turned on causes lag and increases
the latency of turning on the screen. In order to ensure consistent,
maximum performance during/after turning the screen on, use the maximum
bandwidth setting for 10 seconds when the screen is turned on. The
performance impact on screen-wake latency between the minimum bandwidth
vote and the maximum bandwidth vote is significant and measurable, so this
change yields measurable benefits.

Change-Id: I9257fd4a15f3ed8f15b88a5d34b7d94529f0f253
Signed-off-by: Sultanxda <sultanxda@gmail.com>